requirements.txt
mysqlclient==1.4.6
django-crispy-forms==1.8.0